pyproject.toml
[tool.poetry]
name = "elyuf"
version = "0.1.0"
description = "This bot has been created independantly for El-Yurt Umidi Foundation for easily searching Universities global rankings based on three top sources such as QS World University Rank, Times Higher Education Rank, US News & World Report Rank."
authors = ["Mukhsin <muxsinmuxtorov01@gmail.com>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "3.12"
anyio = "4.4.0"
asgiref = "3.8.1"
beautifulsoup4 = "4.12.3"
bs4 = "0.0.2"
certifi = "2024.6.2"
charset-normalizer = "3.3.2"
django = "5.0.6"
h11 = "0.14.0"
httpcore = "1.0.5"
httpx = "0.27.0"
idna = "3.7"
pytelegrambotapi = "4.19.1"
python-dotenv = "1.0.1"
python-telegram-bot = "21.3"
requests = "^2.23.0"
sniffio = "1.3.1"
soupsieve = "2.5"
sqlparse = "0.5.0"
telebot = "0.0.5"
urllib3 = "2.2.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
